fix: add validation  for model endpoint alias
