+++ apps/ai-dial-admin/src/components/EntityView/EntityViewHeaderButtons.tsx
@@ -13,7 +13,7 @@ import AddVersionModal from '@/src/components/Common/AddVersionModal/AddVersionM
 import Button from '@/src/components/Common/Button/Button';
 import ConfirmationModal from '@/src/components/Common/ConfirmationModal/ConfirmationModal';
 import Switch from '@/src/components/Common/Switch/Switch';
-import { isValidEntity } from '@/src/components/EntityListView/CreateEntity/validation';
+import { isValidEntity } from '@/src/components/EntityListView/CreateEntity/utils';
 import { deleteModalTitleMap } from '@/src/components/EntityListView/constants';
 import { getEntityPath } from '@/src/utils/open-in-new-tab';
 import { showEditorErrorNotifications } from '@/src/components/JSONEditor/JSONEditor.utils';
