+++ apps/ai-dial-admin/src/components/ModelView/ModelProperties/ModelTypeProperties.tsx
@@ -8,7 +8,7 @@ import Multiselect from '@/src/components/Common/Multiselect/Multiselect';
 import RadioField from '@/src/components/Common/RadioField/RadioField';
 import EntityAttachments from '@/src/components/EntityView/Properties/EntityAttachments';
 import EntityIcon from '@/src/components/EntityView/Properties/EntityIcon';
-import { EntitiesI18nKey, ModelViewI18nKey, TopicsI18nKey } from '@/src/constants/i18n';
+import { EntitiesI18nKey, ErrorI18nKey, ModelViewI18nKey, TopicsI18nKey } from '@/src/constants/i18n';
 import { RadioFieldOrientation } from '@/src/types/radio-orientation';
 import { getModelsAdapters } from '@/src/app/[lang]/models/actions';
 import { useI18n } from '@/src/locales/client';
@@ -100,6 +100,8 @@ const ModelTypeProperties: FC<Props> = ({ model, onChangeModel }) => {
         postfixPart={postfixPart}
         prefixPart={prefixPart}
         onChange={onChangeEndpoint}
+        invalid={!model.endpointDeploymentName}
+        errorText={t(ErrorI18nKey.IncorrectModelEndpointAlias)}
       />
       <div className="w-full flex flex-col gap-5 lg:w-[35%]">
         <TextInputField
