+++ apps/ai-dial-admin/src/utils/validation/is-valid-adapter.ts
@@ -1,0 +1,12 @@
+import { DialAdapter } from '@/src/models/dial/adapter';
+import { getErrorForDescription } from '@/src/utils/validation/description-error';
+import { getErrorForName } from '@/src/utils/validation/name-error';
+
+export const isValidAdapter = (entity: DialAdapter, names?: string[]) => {
+  return (
+    !!(entity.name && entity.baseEndpoint) &&
+    !getErrorForName(entity.name, names) &&
+    !getErrorForName(entity.displayName) &&
+    !getErrorForDescription(entity.description)
+  );
+};
