+++ apps/ai-dial-admin/src/utils/validation/is-valid-key.ts
@@ -1,0 +1,13 @@
+import { MAX_NAME_SYMBOLS } from '@/src/constants/validation';
+import { DialKey } from '@/src/models/dial/key';
+import { getErrorForDescription } from '@/src/utils/validation/description-error';
+import { getErrorForName } from '@/src/utils/validation/name-error';
+
+export const isValidKey = (entity: DialKey, names?: string[]) => {
+  return (
+    !!(entity.key && entity.project) &&
+    !getErrorForName(entity.name, names) &&
+    !getErrorForDescription(entity.description) &&
+    !(entity.key.length > MAX_NAME_SYMBOLS)
+  );
+};
